# lib/apiservers/portlayer/swagger.json
{
	"swagger": "2.0",
	"basePath": "/",
	"info": {
		"description": "Port Layer API",
		"title": "Port Layer API",
		"version": "v0.0.1"
	},
	"produces": [
		"application/json"
	],
	"consumes": [
		"application/json"
	],
	"schemes": [
		"http"
	],
	"parameters": {
		"operationID": {
			"name": "Op-ID",
			"type": "string",
			"pattern": "[0-9]+\\.[0-9]+",
			"in": "header",
			"required": false
		}
	},
	"paths": {
...
		"/scopes/{idName}": {
			"parameters": [
				{
					"$ref": "#/parameters/operationID"
				}
			],
			"get": {
				"tags": [
					"scopes"
				],
				"operationId": "List",
				"parameters": [
					{
						"name": "idName",
						"type": "string",
						"in": "path",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/ScopeConfig"
							}
						}
					},
					"404": {
						"description": "Not found",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					},
					"default": {
						"description": "error",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					}
				}
			},
			"delete": {
				"tags": [
					"scopes"
				],
				"operationId": "DeleteScope",
				"parameters": [
					{
						"name": "idName",
						"type": "string",
						"in": "path",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"404": {
						"description": "Not found",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					}
				}
			}
		},
		"/scopes/{scope}/containers": {
			"parameters": [
				{
					"$ref": "#/parameters/operationID"
				}
			],
			"post": {
				"description": "Add a container to scopes modifying the container VM's config as necessary",
				"tags": [
					"scopes"
				],
				"operationId": "AddContainer",
				"parameters": [
					{
						"name": "scope",
						"required": true,
						"in": "path",
						"type": "string"
					},
					{
						"name": "config",
						"in": "body",
						"schema": {
							"$ref": "#/definitions/ScopesAddContainerConfig"
						},
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"404": {
						"description": "Not found",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					},
					"500": {
						"description": "error",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					}
				}
			}
		},
		"/scopes/{scope}/containers/{handle}": {
			"parameters": [
				{
					"$ref": "#/parameters/operationID"
				}
			],
			"delete": {
				"description": "Remove a container from a scope",
				"tags": [
					"scopes"
				],
				"operationId": "RemoveContainer",
				"parameters": [
					{
						"name": "handle",
						"required": true,
						"in": "path",
						"type": "string"
					},
					{
						"name": "scope",
						"in": "path",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"404": {
						"description": "Not found",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					},
					"500": {
						"description": "error",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					}
				}
			}
		},
...
		"/scopes/containers/{handle}/binding": {
			"parameters": [
				{
					"$ref": "#/parameters/operationID"
				}
			],
			"post": {
				"tags": [
					"scopes"
				],
				"description": "Perform scope initialization for each of the scopes the container belongs to. Network initialization may include, for example, assigning addresses on a scope.",
				"operationId": "BindContainer",
				"consumes": [
					"application/octet-stream",
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "handle",
						"required": true,
						"in": "path",
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/BindContainerResponse"
						}
					},
					"404": {
						"description": "Not found",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					},
					"500": {
						"description": "error",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					}
				}
			},
			"delete": {
				"tags": [
					"scopes"
				],
				"operationId": "UnbindContainer",
				"consumes": [
					"application/octet-stream",
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "handle",
						"required": true,
						"in": "path",
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/UnbindContainerResponse"
						}
					},
					"404": {
						"description": "Not found",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					},
					"500": {
						"description": "error",
						"schema": {
							"$ref": "#/definitions/Error"
						}
					}
				}
			}
		},
...
	},
	"definitions": {
		"Error": {
			"type": "object",
			"required": [
				"message"
			],
			"properties": {
				"code": {
					"type": "integer",
					"format": "int64"
				},
				"message": {
					"type": "string",
					"x-nullable": false
				}
			}
		},
...
		"ScopeConfig": {
			"type": "object",
			"required": [
				"name",
				"scopeType"
			],
			"properties": {
				"id": {
					"type": "string"
				},
				"name": {
					"x-nullable": false,
					"type": "string"
				},
				"scopeType": {
					"x-nullable": false,
					"type": "string"
				},
				"subnet": {
					"type": "string"
				},
				"gateway": {
					"type": "string"
				},
				"dns": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"ipam": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"endpoints": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/EndpointConfig"
					}
				},
				"annotations": {
					"type": "object",
					"additionalProperties": {
						"type": "string"
					}
				},
				"internal": {
					"type": "boolean"
				}
			}
		},
		"EndpointConfig": {
			"type": "object",
			"required": [
				"id",
				"name",
				"scope",
				"address",
				"gateway",
				"container",
				"ports",
				"aliases",
				"nameservers"
			],
			"properties": {
				"id": {
					"x-nullable": false,
					"type": "string"
				},
				"name": {
					"x-nullable": false,
					"type": "string"
				},
				"scope": {
					"x-nullable": false,
					"type": "string"
				},
				"address": {
					"x-nullable": false,
					"type": "string"
				},
				"gateway": {
					"x-nullable": false,
					"type": "string"
				},
				"container": {
					"x-nullable": false,
					"type": "string"
				},
				"ports": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"aliases": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"nameservers": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"trust": {
					"x-nullable": false,
					"type": "string"
				},
				"direct": {
					"x-nullable": false,
					"type": "boolean"
				},
				"macaddress": {
					"x-nullable": false,
					"type": "string"
				}
			}
		},
...
		"ScopesAddContainerConfig": {
			"type": "object",
			"required": [
				"networkConfig",
				"handle"
			],
			"properties": {
				"networkConfig": {
					"$ref": "#/definitions/NetworkConfig"
				},
				"handle": {
					"x-nullable": false,
					"type": "string"
				}
			}
		},
		"NetworkConfig": {
			"type": "object",
			"required": [
				"networkName"
			],
			"properties": {
				"networkName": {
					"x-nullable": false,
					"type": "string"
				},
				"address": {
					"type": "string"
				},
				"aliases": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"ports": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"nameservers": {
					"type": "array",
					"items": {
						"type": "string"
					}
				}
			}
		},
...
		"BindContainerResponse": {
			"type": "object",
			"required": [
				"handle",
				"endpoints"
			],
			"properties": {
				"handle": {
					"x-nullable": false,
					"type": "string"
				},
				"endpoints": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/EndpointConfig"
					}
				}
			}
		},
		"UnbindContainerResponse": {
			"type": "object",
			"required": [
				"handle",
				"endpoints"
			],
			"properties": {
				"handle": {
					"x-nullable": false,
					"type": "string"
				},
				"endpoints": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/EndpointConfig"
					}
				}
			}
		},
...
	}
}